Added DiskMonitor service
This service is responsible to add an entry in cache which denotes the total disk used up.
Only this entry has non-zero weight in the cache and once the weight of this entry comes
close to cache max-weight, evictions would start.

This benefits us over the old approach of aggressive invalidations of cache entries and re-entries
with the size occupied on disk at that time by lowering the amount of work as invalidation are much
fewer now.
Also, the shift towards executing the old approach in new thread had lead to explosion of threads
slowing down the service.