new:dev:Choose Dropwizard Metrics as Metrics Collection Framework  (#120)
We have wanted to standardize on Dropwizard Metrics for monitoring and the basis of alerting. Rubix depends on Hadoop jars and these bring in their own ecosystem of dependencies and their versions. Therefore we couldn't add Dropwizard Metrics because it may conflict with another version in the classpath. We considered multiple options:
* An interface (#99)
* Shade the jar
* Reuse a version in the class path by using provided scope.

This commit has chosen the last option and has been tested with Presto & Hive.