new:usr: Add a liveness metric for all workers (#152)
Create a heartbeat framework for workers to report liveness to coordinator. The
coordinator can aggregate and send the number of workers to a reporting framework.
This is useful to setup alerts where some RubiX workers maybe non-functional.
The framework will also be used to send other metrics to detect other 
common failures.