Partially filled Blocks should not add weight of full Block to Cache
Last Block of files is usually not aligned to Block boundary, 1MB. But
we were adding the full weight of Block, 1MB, to the Cache for such blocks.
This lead to over-estimation of Cache size and hence early evictions.

With this change, the weight of each file in Cache is represent by it's size in KB
instead of MB. Which means we cannot support file > 1TB in size, which is a big
enough.